requirements.txt
SQLAlchemy==2.0.0
PyMySQL==1.1.0
Flask==3.0.0
Flask-RESTful==0.3.10
dataclasses-json==0.6.2
python-dateutil==2.8.2
